Some work on Captions and Comments

+++ VideoArchiver/App.Domain/Author.cs
@@ -11,7 +11,7 @@ public class Author : BaseArchiveEntity
     public string? DisplayName { get; set; }
     public LangString? Bio { get; set; }
 
-    public int? SubscriberCount { get; set; }
+    public long? SubscriberCount { get; set; }
 
     public List<ImageFile>? ProfileImages { get; set; }
     public List<ImageFile>? Banners { get; set; }
